+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - Added portable Vector class to GridKit
 - Added `REGCA` converter model implementation for PhasorDynamics.
 - Added `GASTPTI` governor model implementation for PhasorDynamics.
+- Added `REECB` converter model implementation for PhasorDynamics.
 
 ## v0.1